Refactor multitask prompt tuning
